Split selectors into different files
