+++ app/javascript/app/components/ghg-emissions/ghg-emissions-selectors/ghg-emissions-selectors-get.js
@@ -1,0 +1,32 @@
+import { createSelector } from 'reselect';
+import { toPlural } from 'utils/ghg-emissions';
+import { generateLinkToDataExplorer } from 'utils/data-explorer';
+
+// meta data for selectors
+export const getData = ({ emissions }) => (emissions && emissions.data) || [];
+export const getMeta = ({ ghgEmissionsMeta }) =>
+  (ghgEmissionsMeta && ghgEmissionsMeta.meta) || null;
+export const getRegions = ({ regions }) => (regions && regions.data) || null;
+export const getSources = createSelector(
+  getMeta,
+  meta => (meta && meta.data_source) || null
+);
+export const getVersions = createSelector(
+  getMeta,
+  meta => (meta && meta.gwp) || null
+);
+export const getWBData = ({ wbCountryData }) => wbCountryData.data || null;
+
+// values from search
+export const getSearch = (state, { search }) => search || null;
+export const getSelection = field =>
+  createSelector(getSearch, search => {
+    if (!search) return null;
+    if (field === 'location') return search.regions || null;
+    return search[field] || search[toPlural(field)] || null;
+  });
+
+export const getLinkToDataExplorer = createSelector([getSearch], search => {
+  const section = 'historical-emissions';
+  return generateLinkToDataExplorer(search, section);
+});
